requirements.txt
fastapi==0.109.2
python-multipart==0.0.7
uvicorn==0.27.1
rembg==2.0.52
Pillow==10.2.0
pillow-heif==0.15.0 